+++ apis/keda/v1alpha1/zz_generated.deepcopy.go
@@ -829,6 +829,11 @@ func (in *ScaledJobSpec) DeepCopyInto(out *ScaledJobSpec) {
 		*out = new(int32)
 		**out = **in
 	}
+	if in.CooldownPeriod != nil {
+		in, out := &in.CooldownPeriod, &out.CooldownPeriod
+		*out = new(int32)
+		**out = **in
+	}
 	in.ScalingStrategy.DeepCopyInto(&out.ScalingStrategy)
 	if in.Triggers != nil {
 		in, out := &in.Triggers, &out.Triggers